+++ src/components/index.js
@@ -1,4 +1,5 @@
-import BookCategory from "./redux/categories/categories";
 import NavBar from "./NavBar";
+import BookCategory from "./BookCategory";
+import BookList from "./BookList";
 
-export {BookCategory, NavBar}
+export {BookCategory, NavBar, BookList}
